[Fix] #32 - split dropbox code
